pyproject.toml
[project]
name = "animate-blueprint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cairosvg>=2.8.2",
    "gitpython>=3.1.45",
    "pillow>=12.0.0",
    "playwright>=1.56.0",
]
